API/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arxiv>=2.4.1",
    "chromadb>=1.5.5",
    "faiss-cpu>=1.13.2",
    "fastapi[standard]>=0.135.2",
    "httpx>=0.28.1",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.22",
    "langchain-groq>=1.1.2",
    "langchain-ollama>=1.0.1",
    "loguru>=0.7.3",
    "orjson>=3.11.7",
    "pydantic>=2.12.5",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "uvicorn>=0.42.0",
    "wikipedia>=1.4.0",
]
